moving files used in Standalone and VstPlugin to Jamtaba-common.pri
