+++ Client/PROJECTS/Jamtaba-common.pri
@@ -46,6 +46,8 @@ HEADERS += NinjamController.h
 HEADERS += ninjam/User.h
 HEADERS += ninjam/Service.h
 HEADERS += ninjam/Server.h
+HEADERS += ninjam/ServerMessages.h
+HEADERS += ninjam/ClientMessages.h
 HEADERS += midi/MidiDriver.h
 HEADERS += gui/plugins/Guis.h
 HEADERS += gui/PluginScanDialog.h
@@ -94,6 +96,8 @@ SOURCES += gui/widgets/WavePeakPanel.cpp
 SOURCES += MainController.cpp
 SOURCES += ninjam/Service.cpp
 SOURCES += ninjam/User.cpp
+SOURCES += ninjam/ServerMessages.cpp
+SOURCES += ninjam/ClientMessages.cpp
 SOURCES += gui/LocalTrackView.cpp
 SOURCES += audio/core/Plugins.cpp
 SOURCES += audio/codec.cpp
