+++ Client/PROJECTS/Standalone/Standalone.pro
@@ -38,8 +38,6 @@ HEADERS += gui/FxPanelItem.h
 HEADERS += audio/core/PortAudioDriver.h
 HEADERS += recorder/JamRecorder.h
 HEADERS += recorder/ReaperProjectGenerator.h
-HEADERS += ninjam/protocol/ServerMessages.h
-HEADERS += ninjam/protocol/ClientMessages.h
 HEADERS += loginserver/natmap.h
 HEADERS += audio/codec.h
 HEADERS += midi/rtMidiDriver.h
@@ -69,8 +67,6 @@ SOURCES += recorder/JamRecorder.cpp
 SOURCES += recorder/ReaperProjectGenerator.cpp
 SOURCES += loginserver/LoginService.cpp
 SOURCES += loginserver/JsonUtils.cpp
-SOURCES += ninjam/protocol/ServerMessages.cpp
-SOURCES += ninjam/protocol/ClientMessages.cpp
 SOURCES += ninjam/Server.cpp
 SOURCES += midi/rtMidiDriver.cpp
 SOURCES += audio/Resampler.cpp
