+++ Client/PROJECTS/VstPlugin/VstPlugin.pro
@@ -44,8 +44,6 @@ DEPENDPATH +=  $$ROOT_PATH/libs/includes/minimp3
 
 HEADERS += recorder/JamRecorder.h
 HEADERS += recorder/ReaperProjectGenerator.h
-HEADERS += ninjam/protocol/ServerMessages.h
-HEADERS += ninjam/protocol/ClientMessages.h
 HEADERS += loginserver/natmap.h
 HEADERS += audio/codec.h
 HEADERS += audio/Resampler.h
@@ -77,8 +75,6 @@ SOURCES += recorder/JamRecorder.cpp
 SOURCES += recorder/ReaperProjectGenerator.cpp
 SOURCES += loginserver/LoginService.cpp
 SOURCES += loginserver/JsonUtils.cpp
-SOURCES += ninjam/protocol/ServerMessages.cpp
-SOURCES += ninjam/protocol/ClientMessages.cpp
 SOURCES += ninjam/Server.cpp
 SOURCES += audio/Resampler.cpp
 SOURCES += audio/vorbis/VorbisDecoder.cpp
